+++ test/integration/world_reset.cc
@@ -27,9 +27,95 @@ typedef std::tr1::tuple<const char *, const char *, int> string2_int;
 class WorldResetTest : public ServerFixture,
                        public ::testing::WithParamInterface<string2_int>
 {
+  /// \brief Test to see if model pose is reset when the world is reset
+  /// \param[in] _physicsEngine Physics engine type.
+  /// \param[in] _world Name of world to load
+  /// \param[in] _resets Number of resets to perform in the test
+  public: void ModelPose(const std::string &_physicsEngine,
+                         const std::string &_world, int _resets);
+
+  /// \brief Test resetting different worlds
+  /// \param[in] _physicsEngine Physics engine type.
+  /// \param[in] _world Name of world to load
+  /// \param[in] _resets Number of resets to perform in the test
   public: void WorldName(const std::string &_physicsEngine,
                          const std::string &_world, int _resets);
 };
+
+/////////////////////////////////////////////////
+void WorldResetTest::ModelPose(const std::string &_physicsEngine,
+                               const std::string &_world, int _resets)
+{
+  if (_physicsEngine == "dart")
+  {
+    gzerr << "Abort test since dart does not support ray sensor in PR2, "
+          << "Please see issue #911.\n";
+    return;
+  }
+
+  Load(_world, true, _physicsEngine);
+  physics::WorldPtr world = physics::get_world("default");
+  ASSERT_TRUE(world != NULL);
+
+  physics::PhysicsEnginePtr physics = world->GetPhysicsEngine();
+  ASSERT_TRUE(physics != NULL);
+  EXPECT_EQ(physics->GetType(), _physicsEngine);
+
+  double dt = physics->GetMaxStepSize();
+  unsigned int steps = 250;
+
+  // Step forward, verify time increasing
+  world->Step(steps);
+  double simTime = world->GetSimTime().Double();
+  EXPECT_NEAR(simTime, dt*steps, dt);
+
+  ignition::math::Pose3d initialPose(1, 2, 0.5, 0, 0, 1.57);
+
+  // spawn a box with known initial pose
+  math::Vector3 size(1, 1, 1);
+  SpawnBox("box", size, initialPose.Pos(), initialPose.Rot().Euler(), false);
+  physics::ModelPtr model = world->GetModel("box");
+  ASSERT_TRUE(model != NULL);
+
+  // verify box pose
+  EXPECT_EQ(model->GetWorldPose(), initialPose);
+
+  // move box to new pose
+  ignition::math::Pose3d newPose(4, 5, 0.5, 0, 0, 0);
+  model->SetWorldPose(newPose);
+  EXPECT_EQ(model->GetWorldPose(), newPose);
+
+  // Reset world repeatedly
+  for (int i = 0; i < _resets; ++i)
+  {
+    // Reset world, verify time == 0
+    world->Reset();
+    simTime = world->GetSimTime().Double();
+    EXPECT_NEAR(simTime, 0.0, dt);
+
+    // Step forward, verify time increasing
+    world->Step(steps);
+    simTime = world->GetSimTime().Double();
+    EXPECT_NEAR(simTime, dt*steps, dt);
+  }
+
+  // verify box has moved back to initial pose
+  EXPECT_EQ(model->GetWorldPose(), initialPose);
+}
+
+/////////////////////////////////////////////////
+TEST_P(WorldResetTest, ModelPose)
+{
+  std::string physics = std::tr1::get<0>(GetParam());
+
+  std::string worldName = std::tr1::get<1>(GetParam());
+  int resets = std::tr1::get<2>(GetParam());
+  gzdbg << "Physics engine [" << physics << "] "
+        << "world name [" << worldName << "] "
+        << "reset count [" << resets << "]"
+        << std::endl;
+  ModelPose(physics, worldName, resets);
+}
 
 /////////////////////////////////////////////////
 void WorldResetTest::WorldName(const std::string &_physicsEngine,